requirements.txt
pypdf
langchain
langchain-community
chromadb # Vector storage
pytest
boto3
streamlit
python-dotenv
langchain-openai